ophthalmology_llm/requirements.txt
torch
torchvision
pillow
tqdm
pandas
numpy
scikit-learn
matplotlib
